requirements.txt
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
Deprecated==1.2.13
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.92.0
fastapi-users==10.4.0
fastapi-users-db-sqlalchemy==5.0.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
makefun==1.15.1
opentelemetry-api==1.16.0
opentelemetry-distro==0.37b0
opentelemetry-instrumentation==0.37b0
opentelemetry-sdk==1.16.0
opentelemetry-semantic-conventions==0.37b0
passlib==1.7.4
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-dotenv==1.0.0
python-multipart==0.0.5
sentry-sdk==1.16.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
wrapt==1.15.0

openai~=0.27.2
Pillow~=9.4.0